fix: warning

+++ src/02_a_basic_calculator/10_conversions_as_casting/src/lib.rs
@@ -19,7 +19,7 @@ mod tests {
         // literal. If we were to use a variable, the compiler wouldn't be able to
         // catch this at compile time.
         #[allow(overflowing_literals)]
-        let x = { 255 as i8 };
+        let x = { 255_i8 };
 
         // You could solve this by using exactly the same expression as above,
         // but that would defeat the purpose of the exercise. Instead, use a genuine
